day-3/web-api/requirements.txt
fastapi
tensorflow
numpy
pillow
uvicorn
huggingface_hub
python-multipart